Add support for Neovim
Although the current plugin uses the function `term_start` to run a
command, this function isn't available on Neovim.

Changed to use the function `termopen` instead if an editor is Neovim,
otherwise `term_start` is called as before.  To disable unnecessary
prints in the Neovim terminal, some local options are changed.  Also, as
the `termopen` opens a new terminal in the current buffer, `wincmd` is
called in advance to split a window vertically.

Signed-off-by: h1r0mu <4m3on0@gmail.com>